requirements.txt
groq
requests
beautifulsoup4
python-dotenv